backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
onnxruntime
numpy
pillow
python-multipart